mejora codigo y quita combo camaras
