+++ src/main/java/fxml/controller/WebCamPreviewController.java
@@ -13,9 +13,9 @@ import org.apache.commons.lang3.StringUtils;
 
 import com.github.sarxos.webcam.Webcam;
 
+import fxml.entity.WebCamInfo;
 import javafx.application.Platform;
 import javafx.beans.property.ObjectProperty;
-import javafx.stage.Stage;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.value.ChangeListener;
 import javafx.beans.value.ObservableValue;
@@ -27,11 +27,11 @@ import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
 import javafx.fxml.Initializable;
 import javafx.scene.control.Button;
-import javafx.scene.control.ComboBox;
 import javafx.scene.control.Slider;
 import javafx.scene.image.Image;
 import javafx.scene.image.ImageView;
 import javafx.scene.layout.BorderPane;
+import javafx.stage.Stage;
 import laucher.AppLauncher;
 
 @SuppressWarnings("restriction")
@@ -78,8 +78,6 @@ public class WebCamPreviewController implements Initializable {
 	@FXML
 	Button btnDisposeCamera;
 	@FXML
-	ComboBox<WebCamInfo> cbCameraOptions;
-	@FXML
 	BorderPane bpWebCamPaneHolder;
 	@FXML
 	ImageView imgWebCamCapturedImage;
@@ -114,7 +112,7 @@ public class WebCamPreviewController implements Initializable {
 
 	public BufferedImage[] lBuffered;
 	public static int speed = 30;
-
+	public ObservableList<WebCamInfo> options;
 	public ArrayDeque<BufferedImage> aqImagenes;
 	public ThreadSlider tslider;
 
@@ -127,7 +125,6 @@ public class WebCamPreviewController implements Initializable {
 	private BufferedImage grabbedImage;
 
 	private CapturedImage[] arrayImg;
-	// private WebcamPanel selWebCamPanel = null;
 	private Webcam selWebCam = null;
 
 	private List<javafx.scene.control.RadioButton> radioList = new ArrayList<javafx.scene.control.RadioButton>();
@@ -139,8 +136,6 @@ public class WebCamPreviewController implements Initializable {
 	private ObjectProperty<Image> imageProperty3 = new SimpleObjectProperty<Image>();
 	private ObjectProperty<Image> miniFramePreview = new SimpleObjectProperty<Image>();
 
-	private String cameraListPromptText = "Seleccion Camara";
-
 	public javafx.scene.control.Label getLabelFPS() {
 		return labelFPS;
 	}
@@ -149,18 +144,14 @@ public class WebCamPreviewController implements Initializable {
 		labelFPS.setText(value);
 	}
 
-	// @Override
 	public void initialize(URL arg0, ResourceBundle arg1) {
-
 		this.cmbConfig.setItems(WindowUtil.getConfig());
 		cmbConfig.getSelectionModel().selectFirst();
-
 		radioList.add(radioImg1);
 		radioList.add(radioImg2);
 		radioList.add(radioImg3);
 		radioList.add(radioImg4);
-		// fpBottomPane.setDisable(true);
-		ObservableList<WebCamInfo> options = FXCollections.observableArrayList();
+		options = FXCollections.observableArrayList();
 		int webCamCounter = 0;
 		for (Webcam webcam : Webcam.getWebcams()) {
 			WebCamInfo webCamInfo = new WebCamInfo();
@@ -170,21 +161,6 @@ public class WebCamPreviewController implements Initializable {
 			options.add(webCamInfo);
 			webCamCounter++;
 		}
-		cbCameraOptions.setItems(options);
-		// cbCameraOptions.getSelectionModel().select(0);
-		// initializeWebCam(webCamIndex);
-		cbCameraOptions.setPromptText(cameraListPromptText);
-		cbCameraOptions.getSelectionModel().selectedItemProperty().addListener(new ChangeListener<WebCamInfo>() {
-
-			// @Override
-			public void changed(ObservableValue<? extends WebCamInfo> arg0, WebCamInfo arg1, WebCamInfo arg2) {
-				if (arg2 != null) {
-					System.out.println(
-							"WebCam Index: " + arg2.getWebCamIndex() + ": WebCam Name:" + arg2.getWebCamName());
-					initializeWebCam(arg2.getWebCamIndex());
-				}
-			}
-		});
 
 		Platform.runLater(new Runnable() {
 			// @Override
@@ -196,11 +172,9 @@ public class WebCamPreviewController implements Initializable {
 		checkFilter1.selectedProperty().addListener(new ChangeListener<Boolean>() {
 			@Override
 			public void changed(ObservableValue<? extends Boolean> observable, Boolean oldValue, Boolean newValue) {
-//		        chk2.setSelected(!newValue);
 				WebCamPreviewController.filter1Active = !WebCamPreviewController.filter1Active;
 			}
 		});
-
 	}
 
 	protected void setImageViewSize() {
@@ -229,12 +203,9 @@ public class WebCamPreviewController implements Initializable {
 	}
 
 	protected void initializeWebCam(final int webCamIndex) {
-
 		Task<Void> webCamIntilizer = new Task<Void>() {
-
 			@Override
 			protected Void call() throws Exception {
-
 				if (selWebCam == null) {
 					selWebCam = Webcam.getWebcams().get(webCamIndex);
 					selWebCam.open();
@@ -242,26 +213,18 @@ public class WebCamPreviewController implements Initializable {
 					closeCamera();
 					selWebCam = Webcam.getWebcams().get(webCamIndex);
 					selWebCam.open();
-
 				}
 				startWebCamStream();
 				return null;
 			}
-
 		};
-
 		new Thread(webCamIntilizer).start();
-		// fpBottomPane.setDisable(false);
 		btnStartCamera.setDisable(true);
-
-		System.out.println("this.cmbConfig.getSelectionModel().getSelectedItem() "
-				+ this.cmbConfig.getSelectionModel().getSelectedItem());
 		windows = new WindowUtil().openWindows(imgWebCamCapturedImage2, imgWebCamCapturedImage3,
 				imgWebCamCapturedImage4, imgWebCamCapturedImage5, this.cmbConfig.getSelectionModel().getSelectedItem());
 	}
 
 	protected void startWebCamStream() {
-
 		stopCamera = false;
 		task = new TaskCamera(stopCamera, grabbedImage, webcamDefault, arrayImg, imageProperty, imageProperty2,
 				imageProperty3, imageProperty3, miniFramePreview, this);
@@ -299,49 +262,16 @@ public class WebCamPreviewController implements Initializable {
 
 	public void startCamera(ActionEvent event) {
 //		stopCamera = false;
-		task.stopCamera = false; // funciona este
-		tslider.cancel();
+		if (task != null) {
+			task.stopCamera = false; // funciona este
+			tslider.cancel();
+		}
+		initializeWebCam(0);
 		startWebCamStream();
-
 		btnStartCamera.setDisable(true);
 		btnStopCamera.setDisable(false);
-
 	}
 
-	// public void disposeCamera(ActionEvent event) {
-	// stopCamera = true;
-	// closeCamera();
-	// Webcam.shutdown();
-	// btnStopCamera.setDisable(true);
-	// btnStartCamera.setDisable(true);
-	// }
-
-	class WebCamInfo {
-		private String webCamName;
-		private int webCamIndex;
-
-		public String getWebCamName() {
-			return webCamName;
-		}
-
-		public void setWebCamName(String webCamName) {
-			this.webCamName = webCamName;
-		}
-
-		public int getWebCamIndex() {
-			return webCamIndex;
-		}
-
-		public void setWebCamIndex(int webCamIndex) {
-			this.webCamIndex = webCamIndex;
-		}
-
-		@Override
-		public String toString() {
-			return webCamName;
-		}
-
-	}
 
 	public void createSlider(final ArrayDeque<BufferedImage> imagenes) {
 		// play images
