+++ src/main/java/fxml/gui/WebCamPreview2.fxml
@@ -33,15 +33,6 @@
                 <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
               </rowConstraints>
                <children>
-                  <ComboBox fx:id="cbCameraOptions" prefHeight="30.0" prefWidth="140.0" style="-fx-font-size: 12px;" GridPane.columnIndex="1" GridPane.halignment="LEFT" GridPane.valignment="CENTER" />
-              <Label text="Camara" GridPane.halignment="RIGHT" GridPane.hgrow="ALWAYS" GridPane.valignment="CENTER" GridPane.vgrow="ALWAYS">
-                <GridPane.margin>
-                  <Insets left="5.0" right="10.0" />
-                </GridPane.margin>
-                     <font>
-                        <Font name="Segoe UI" size="12.0" fx:id="x4" />
-                     </font>
-              </Label>
                   <Label alignment="CENTER" text="FPS" GridPane.halignment="RIGHT" GridPane.rowIndex="1" GridPane.valignment="CENTER">
                      <font>
                         <Font name="Segoe UI" size="12.0" />
@@ -52,22 +43,7 @@
                   </Label>
                   <Label fx:id="labelFPS" layoutX="142.0" layoutY="47.0" />
                   <Label fx:id="labelFPS" prefHeight="18.0" prefWidth="90.0" GridPane.columnIndex="1" GridPane.rowIndex="1" />
-               </children>
-               <GridPane.margin>
-                  <Insets right="10.0" />
-               </GridPane.margin>
-            </GridPane>
-            <GridPane prefHeight="60.0" prefWidth="252.0" GridPane.columnIndex="1">
-              <columnConstraints>
-                <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-                <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-              </columnConstraints>
-              <rowConstraints>
-                <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-                <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-              </rowConstraints>
-               <children>
-              <Button fx:id="btnStartCamera" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" onAction="#startCamera" prefHeight="28.0" prefWidth="78.0" text="Grabar">
+              <Button fx:id="btnStartCamera" contentDisplay="CENTER" minHeight="-Infinity" minWidth="-Infinity" mnemonicParsing="false" onAction="#startCamera" prefHeight="28.0" prefWidth="78.0" text="Grabar" textAlignment="CENTER" GridPane.halignment="RIGHT">
                 <font>
                   <Font name="Segoe UI" size="12.0" fx:id="x1" />
                 </font>
@@ -84,6 +60,19 @@
                      </GridPane.margin>
                   </Button>
                </children>
+               <GridPane.margin>
+                  <Insets right="10.0" />
+               </GridPane.margin>
+            </GridPane>
+            <GridPane prefHeight="60.0" prefWidth="252.0" GridPane.columnIndex="1">
+              <columnConstraints>
+                <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+                <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
+              </columnConstraints>
+              <rowConstraints>
+                <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+                <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+              </rowConstraints>
             </GridPane>
             <GridPane prefHeight="80.0" prefWidth="751.0" GridPane.columnIndex="2">
               <columnConstraints>
